Experimenting with FlattenLayers
A summary of all changes from this commit and parent:

- Added an optional argument to the FlattenLayers constructor to
trigger the behaviour specific to merging down when it executes.
Mostly it just sets the bottom layer as destination, and then
ensures notifyLayerMergedDown will be called before removing the
source layers. Packing flags into an int would be a better idea,
but I didn't want to touch the function signature too much.

- Calling shrink_bounds at the top of the loop, rather than
conditionally, and with the additional argument to then crop the
rendered image if it's not fully transparent. Makes it so the layer
edges display correctly.

- If the bottom cel has a z-index greater than zero, then it
accounts for the top layer being removed by decreasing it's
z-index. This helps stop the result from 'popping' through cels
that used to be drawn on top of it, but it isn't perfect.
Inspecting the "surrounding" layers before doing any such
adjustment to make sure it won't *visually* modify the final image
would be more accurate.

- The destination layer's blend mode is reset to normal if no new
flat layer is made, and when layer/cel opacity is less than 100%,
the result doesn't become more transparent. No notes, it's just two
things that bugged me a little.