Added basic 'AdjustZIndex' on merge
Made this functionality optional; it works in some cases, and in others it doesn't. It's tough to accurately predict how changes to the z-index would modify the image, and there are instances where nothing can really be done to prevent it. Nevertheless, this solves the few viable cases I had in mind.